Move terra-library folder to lib/
