+++ packages/client/src/modules/ledger/epics.ts
@@ -7,6 +7,7 @@ import {
   wait,
 } from "@anthem/utils";
 import { LEDGER_ERRORS } from "constants/ledger-errors";
+import { getAccAddress } from "lib/terra-lib/key-utils";
 import { EpicSignature, ReduxActionTypes } from "modules/root";
 import { i18nSelector } from "modules/settings/selectors";
 import { combineEpics } from "redux-observable";
@@ -29,7 +30,6 @@ import {
   onPageWhichIncludesAddressParam,
 } from "tools/client-utils";
 import StorageModule from "tools/storage-utils";
-import { getAccAddress } from "tools/terra-library/key-utils";
 import {
   validateLedgerAppVersion,
   validateNetworkAddress,
